Offline - Better clean architecture layer separation
Separated api response models from datasource models
Created response->datasource mappers
Network datasource implementaiton
