+++ app/src/main/java/dev/jameido/pokedex/PokedexApplication.kt
@@ -1,10 +1,7 @@
 package dev.jameido.pokedex
 
 import android.app.Application
-import dev.jameido.pokedex.di.repositoryModule
-import dev.jameido.pokedex.di.retrofitModule
-import dev.jameido.pokedex.di.useCaseModule
-import dev.jameido.pokedex.di.viewModelModule
+import dev.jameido.pokedex.di.*
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
 
@@ -19,6 +16,7 @@ class PokedexApplication : Application() {
             androidContext(this@PokedexApplication)
             modules(listOf(
                     retrofitModule,
+                    dataSourceModule,
                     repositoryModule,
                     useCaseModule,
                     viewModelModule
