+++ app/src/main/java/dev/jameido/pokedex/data/datasource/PkmnDataSource.kt
@@ -1,18 +1,18 @@
 package dev.jameido.pokedex.data.datasource
 
-import dev.jameido.pokedex.data.models.ResPkmnDetail
-import dev.jameido.pokedex.data.models.ResPkmnList
-import dev.jameido.pokedex.data.models.ResPkmnSpecies
+import dev.jameido.pokedex.data.models.PkmnDetailModel
+import dev.jameido.pokedex.data.models.PkmnListModel
+import dev.jameido.pokedex.data.models.PkmnSpeciesModel
 
 /**
  * Created by Jameido on 19/12/2020.
  */
 interface PkmnDataSource {
-    suspend fun list(limit: Int, offset: Int) : ResPkmnList
+    suspend fun list(pageSize: Int, page: Int) : PkmnListModel
 
-    suspend fun detail(name: String) : ResPkmnDetail
+    suspend fun detail(name: String) : PkmnDetailModel
 
-    suspend fun species(name: String) : ResPkmnSpecies
+    suspend fun species(name: String) : PkmnSpeciesModel
 }
 
 interface NetworkPkmnDataSource : PkmnDataSource
