+++ app/src/main/java/dev/jameido/pokedex/data/mappers/PkmnEntityMapper.kt
@@ -1,14 +1,14 @@
 package dev.jameido.pokedex.data.mappers
 
 import android.net.Uri
-import dev.jameido.pokedex.data.models.PkmnElement
+import dev.jameido.pokedex.data.models.PkmnModel
 import dev.jameido.pokedex.domain.entity.PkmnEntity
 
 /**
  * Created by Jameido on 04/01/2021.
  */
-class PkmnMapper : EntityMapper<PkmnElement, PkmnEntity> {
-    override fun map(model: PkmnElement): PkmnEntity {
+class PkmnEntityMapper : EntityMapper<PkmnModel, PkmnEntity> {
+    override fun map(model: PkmnModel): PkmnEntity {
         var index: Int? = null
         var sprite: String? = null
         model.url?.let { url ->
