+++ app/src/main/java/dev/jameido/pokedex/data/models/PkmnDetailModel.kt
@@ -1,0 +1,19 @@
+package dev.jameido.pokedex.data.models
+
+/**
+ * Created by Jameido on 20/12/2020.
+ */
+data class PkmnDetailModel(
+        val id: Int,
+        val name: String,
+        val height: Float,
+        val weight: Float,
+        val sprite: String?,
+        val stats: List<Stat>,
+        val types: List<String>
+)
+
+data class Stat(
+        val value: Int,
+        val name: String
+)
