--- app/src/main/java/dev/jameido/pokedex/data/models/PkmnElement.kt
@@ -1,6 +1,0 @@
-package dev.jameido.pokedex.data.models
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class PkmnElement(val name: String = "", val url: String?)
