+++ app/src/main/java/dev/jameido/pokedex/data/models/PkmnListModel.kt
@@ -1,0 +1,6 @@
+package dev.jameido.pokedex.data.models
+
+/**
+ * Created by Jameido on 04/01/2021.
+ */
+data class PkmnListModel(val next: Int?, val previous: Int?, val results: List<PkmnModel>)
