+++ app/src/main/java/dev/jameido/pokedex/data/models/PkmnModel.kt
@@ -1,0 +1,6 @@
+package dev.jameido.pokedex.data.models
+
+/**
+ * Created by Jameido on 04/01/2021.
+ */
+data class PkmnModel(val name: String = "", val url: String?)
