+++ app/src/main/java/dev/jameido/pokedex/framework/datasource/network/mappers/PkmnSpeciesMapper.kt
@@ -1,0 +1,16 @@
+package dev.jameido.pokedex.framework.datasource.network.mappers
+import dev.jameido.pokedex.data.models.PkmnSpeciesModel
+import dev.jameido.pokedex.framework.datasource.network.models.ResPkmnSpecies
+
+/**
+ * Created by Jameido on 04/01/2021.
+ */
+class PkmnSpeciesMapper : ResponseMapper<ResPkmnSpecies, PkmnSpeciesModel> {
+
+    override fun map(model: ResPkmnSpecies): PkmnSpeciesModel {
+        val pkmnMapper = PkmnMapper()
+        val varieties = model.varieties.map { pkmnMapper.map(it.pokemon) }
+        val description = model.flavor_text_entries.first { it.language.name == "en" }.flavor_text
+        return PkmnSpeciesModel(model.id, model.name, description, varieties, model.evolution_chain.url)
+    }
+}
