+++ app/src/main/java/dev/jameido/pokedex/framework/datasource/network/models/ResPkmnElement.kt
@@ -1,0 +1,6 @@
+package dev.jameido.pokedex.framework.datasource.network.models
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class ResPkmnElement(val name: String = "", val url: String?)
