+++ app/src/test/java/dev/jameido/pokedex/framework/ApiUnitTest.kt
@@ -1,46 +1,123 @@
-package dev.jameido.pokedex.data.models
+package dev.jameido.pokedex.framework
 
-import com.squareup.moshi.JsonDataException
-import com.squareup.moshi.Moshi
+import dev.jameido.pokedex.framework.datasource.network.PkmnApi
+import kotlinx.coroutines.runBlocking
+import okhttp3.mockwebserver.MockResponse
+import okhttp3.mockwebserver.MockWebServer
+import org.junit.After
 import org.junit.Assert.*
+import org.junit.Before
 import org.junit.Test
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
 
 /**
- * Created by Jameido on 02/01/2021.
+ * Created by Jameido on 18/12/2020.
  */
-class PkmnSpeciesUnitTest {
-    private val moshi = Moshi.Builder()
-            .build()
+class ApiUnitTest {
+
+    private lateinit var server: MockWebServer
+    private lateinit var mockApi: PkmnApi
 
     @Test
-    fun parsePkmnDetail() {
-        try {
-            val detailResponse = moshi.adapter(ResPkmnSpecies::class.java).fromJson(DETAIL_CHARIZARD)
-            assertNotNull(detailResponse)
-            assertEquals(detailResponse!!.id, 6)
-            assertEquals(detailResponse!!.name, "charizard")
-            assertNotNull(detailResponse!!.evolution_chain)
-            assertEquals(detailResponse!!.evolution_chain.url, "https://pokeapi.co/api/v2/evolution-chain/2/")
-            assertEquals(detailResponse!!.flavor_text_entries.size, 154)
-            assertNotNull(detailResponse!!.flavor_text_entries[0])
-            // \f has been replaced with u000C to avoid compilation issues, as reported at https://kotlinlang.org/docs/reference/basic-types.html#characters
-            // The following escape sequences are supported: \t, \b, \n, \r, \', \", \\ and \$
-            assertEquals(detailResponse!!.flavor_text_entries[0].flavor_text, "Spits fire that\nis hot enough to\nmelt boulders.\u000CKnown to cause\nforest fires\nunintentionally.")
-            assertNotNull(detailResponse!!.flavor_text_entries[0].language)
-            assertEquals(detailResponse!!.flavor_text_entries[0].language.name, "en")
-            assertEquals(detailResponse!!.varieties.size, 4)
-            assertNotNull(detailResponse!!.varieties[0])
-            assertEquals(detailResponse!!.varieties[0].is_default, true)
-            assertNotNull(detailResponse!!.varieties[0].pokemon)
-            assertEquals(detailResponse!!.varieties[0].pokemon.url, "https://pokeapi.co/api/v2/pokemon/6/")
-            assertEquals(detailResponse!!.varieties[0].pokemon.name, "charizard")
+    fun getPkmnList() {
 
-        } catch (ex: JsonDataException) {
-            fail()
+        server.enqueue(MockResponse().setBody(LIST_FIRST_ELEMENTS))
+        val response = runBlocking {
+            mockApi.list(0, 20)
         }
+
+        assertNotNull(response)
+        assertEquals(response.count, 898)
+        assertEquals(response.next, "https://pokeapi.co/api/v2/pokemon-species/?offset=20&limit=20")
+        assertNull(response.previous)
+        assertNotNull(response.results)
+        assertEquals(response.results.size, 20)
+        assertEquals(response.results[0], PkmnElement("bulbasaur", "https://pokeapi.co/api/v2/pokemon-species/1/"))
+
+    }
+    @Test
+    fun getPkmnDetail() {
+
+        server.enqueue(MockResponse().setBody(DETAIL_DITTO))
+        val response = runBlocking {
+            mockApi.detail("ditto")
+        }
+
+        assertNotNull(response)
+        assertEquals(response!!.id, 132)
+        assertEquals(response!!.name, "ditto")
+        assertNotNull(response!!.sprites)
+        assertEquals(response!!.sprites.front_default, "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/132.png")
+        assertEquals(response!!.stats.size, 6)
+        assertNotNull(response!!.stats[0])
+        assertEquals(response!!.stats[0].base_stat, 48)
+        assertNotNull(response!!.types[0].type)
+        assertEquals(response!!.types[0].type.name, "normal")
+    }
+    @Test
+    fun getPkmnSpecies() {
+
+        server.enqueue(MockResponse().setBody(DETAIL_CHARIZARD))
+        val response = runBlocking {
+            mockApi.species("charizard")
+        }
+
+        assertNotNull(response)
+        assertEquals(response!!.id, 6)
+        assertEquals(response!!.name, "charizard")
+        assertNotNull(response!!.evolution_chain)
+        assertEquals(response!!.evolution_chain.url, "https://pokeapi.co/api/v2/evolution-chain/2/")
+        assertEquals(response!!.flavor_text_entries.size, 154)
+        assertNotNull(response!!.flavor_text_entries[0])
+        // \f has been replaced with u000C to avoid compilation issues, as reported at https://kotlinlang.org/docs/reference/basic-types.html#characters
+        // The following escape sequences are supported: \t, \b, \n, \r, \', \", \\ and \$
+        assertEquals(response!!.flavor_text_entries[0].flavor_text, "Spits fire that\nis hot enough to\nmelt boulders.\u000CKnown to cause\nforest fires\nunintentionally.")
+        assertNotNull(response!!.flavor_text_entries[0].language)
+        assertEquals(response!!.flavor_text_entries[0].language.name, "en")
+        assertEquals(response!!.varieties.size, 4)
+        assertNotNull(response!!.varieties[0])
+        assertEquals(response!!.varieties[0].is_default, true)
+        assertNotNull(response!!.varieties[0].pokemon)
+        assertEquals(response!!.varieties[0].pokemon.url, "https://pokeapi.co/api/v2/pokemon/6/")
+        assertEquals(response!!.varieties[0].pokemon.name, "charizard")
+
+    }
+
+    @Before
+    fun setUp() {
+        server = MockWebServer().apply {
+            start()
+        }
+        mockApi = Retrofit.Builder()
+                .baseUrl(server.url("/"))
+                .addConverterFactory(MoshiConverterFactory.create())
+                .build()
+                .create(PkmnApi::class.java)
+    }
+
+    @After
+    fun tearDown() {
+        server.shutdown()
     }
 
     companion object {
+        //Response of pokemon-species?limit=20&offset=0
+        const val LIST_FIRST_ELEMENTS = """{"count":898,"next":"https://pokeapi.co/api/v2/pokemon-species/?offset=20&limit=20","previous":null,"results":[{"name":"bulbasaur","url":"https://pokeapi.co/api/v2/pokemon-species/1/"},{"name":"ivysaur","url":"https://pokeapi.co/api/v2/pokemon-species/2/"},{"name":"venusaur","url":"https://pokeapi.co/api/v2/pokemon-species/3/"},{"name":"charmander","url":"https://pokeapi.co/api/v2/pokemon-species/4/"},{"name":"charmeleon","url":"https://pokeapi.co/api/v2/pokemon-species/5/"},{"name":"charizard","url":"https://pokeapi.co/api/v2/pokemon-species/6/"},{"name":"squirtle","url":"https://pokeapi.co/api/v2/pokemon-species/7/"},{"name":"wartortle","url":"https://pokeapi.co/api/v2/pokemon-species/8/"},{"name":"blastoise","url":"https://pokeapi.co/api/v2/pokemon-species/9/"},{"name":"caterpie","url":"https://pokeapi.co/api/v2/pokemon-species/10/"},{"name":"metapod","url":"https://pokeapi.co/api/v2/pokemon-species/11/"},{"name":"butterfree","url":"https://pokeapi.co/api/v2/pokemon-species/12/"},{"name":"weedle","url":"https://pokeapi.co/api/v2/pokemon-species/13/"},{"name":"kakuna","url":"https://pokeapi.co/api/v2/pokemon-species/14/"},{"name":"beedrill","url":"https://pokeapi.co/api/v2/pokemon-species/15/"},{"name":"pidgey","url":"https://pokeapi.co/api/v2/pokemon-species/16/"},{"name":"pidgeotto","url":"https://pokeapi.co/api/v2/pokemon-species/17/"},{"name":"pidgeot","url":"https://pokeapi.co/api/v2/pokemon-species/18/"},{"name":"rattata","url":"https://pokeapi.co/api/v2/pokemon-species/19/"},{"name":"raticate","url":"https://pokeapi.co/api/v2/pokemon-species/20/"}]}"""
+
+        //Response of pokemon-species?limit=20&offset=880
+        const val LIST_LAST_ELEMENTS = """{"count":898,"next":null,"previous":"https://pokeapi.co/api/v2/pokemon-species/?offset=860&limit=20","results":[{"name":"arctozolt","url":"https://pokeapi.co/api/v2/pokemon-species/881/"},{"name":"dracovish","url":"https://pokeapi.co/api/v2/pokemon-species/882/"},{"name":"arctovish","url":"https://pokeapi.co/api/v2/pokemon-species/883/"},{"name":"duraludon","url":"https://pokeapi.co/api/v2/pokemon-species/884/"},{"name":"dreepy","url":"https://pokeapi.co/api/v2/pokemon-species/885/"},{"name":"drakloak","url":"https://pokeapi.co/api/v2/pokemon-species/886/"},{"name":"dragapult","url":"https://pokeapi.co/api/v2/pokemon-species/887/"},{"name":"zacian","url":"https://pokeapi.co/api/v2/pokemon-species/888/"},{"name":"zamazenta","url":"https://pokeapi.co/api/v2/pokemon-species/889/"},{"name":"eternatus","url":"https://pokeapi.co/api/v2/pokemon-species/890/"},{"name":"kubfu","url":"https://pokeapi.co/api/v2/pokemon-species/891/"},{"name":"urshifu","url":"https://pokeapi.co/api/v2/pokemon-species/892/"},{"name":"zarude","url":"https://pokeapi.co/api/v2/pokemon-species/893/"},{"name":"regieleki","url":"https://pokeapi.co/api/v2/pokemon-species/894/"},{"name":"regidrago","url":"https://pokeapi.co/api/v2/pokemon-species/895/"},{"name":"glastrier","url":"https://pokeapi.co/api/v2/pokemon-species/896/"},{"name":"spectrier","url":"https://pokeapi.co/api/v2/pokemon-species/897/"},{"name":"calyrex","url":"https://pokeapi.co/api/v2/pokemon-species/898/"}]}"""
+
+        //Response of pokemon-species?limit=20&offset=500
+        const val LIST_INTERMEDIATE = """{"count":898,"next":"https://pokeapi.co/api/v2/pokemon-species/?offset=520&limit=20","previous":"https://pokeapi.co/api/v2/pokemon-species/?offset=480&limit=20","results":[{"name":"oshawott","url":"https://pokeapi.co/api/v2/pokemon-species/501/"},{"name":"dewott","url":"https://pokeapi.co/api/v2/pokemon-species/502/"},{"name":"samurott","url":"https://pokeapi.co/api/v2/pokemon-species/503/"},{"name":"patrat","url":"https://pokeapi.co/api/v2/pokemon-species/504/"},{"name":"watchog","url":"https://pokeapi.co/api/v2/pokemon-species/505/"},{"name":"lillipup","url":"https://pokeapi.co/api/v2/pokemon-species/506/"},{"name":"herdier","url":"https://pokeapi.co/api/v2/pokemon-species/507/"},{"name":"stoutland","url":"https://pokeapi.co/api/v2/pokemon-species/508/"},{"name":"purrloin","url":"https://pokeapi.co/api/v2/pokemon-species/509/"},{"name":"liepard","url":"https://pokeapi.co/api/v2/pokemon-species/510/"},{"name":"pansage","url":"https://pokeapi.co/api/v2/pokemon-species/511/"},{"name":"simisage","url":"https://pokeapi.co/api/v2/pokemon-species/512/"},{"name":"pansear","url":"https://pokeapi.co/api/v2/pokemon-species/513/"},{"name":"simisear","url":"https://pokeapi.co/api/v2/pokemon-species/514/"},{"name":"panpour","url":"https://pokeapi.co/api/v2/pokemon-species/515/"},{"name":"simipour","url":"https://pokeapi.co/api/v2/pokemon-species/516/"},{"name":"munna","url":"https://pokeapi.co/api/v2/pokemon-species/517/"},{"name":"musharna","url":"https://pokeapi.co/api/v2/pokemon-species/518/"},{"name":"pidove","url":"https://pokeapi.co/api/v2/pokemon-species/519/"},{"name":"tranquill","url":"https://pokeapi.co/api/v2/pokemon-species/520/"}]}"""
+
+        //Response of pokemon-species?limit=20&offset=2000
+        const val LIST_OUT_BOUNDS = """{"count":898,"next":null,"previous":"https://pokeapi.co/api/v2/pokemon-species/?offset=1980&limit=20","results":[]}"""
+
+        //Response of pokemon/ditto or pokemon/132
+        const val DETAIL_DITTO = """{"abilities":[{"ability":{"name":"limber","url":"https://pokeapi.co/api/v2/ability/7/"},"is_hidden":false,"slot":1},{"ability":{"name":"imposter","url":"https://pokeapi.co/api/v2/ability/150/"},"is_hidden":true,"slot":3}],"base_experience":101,"forms":[{"name":"ditto","url":"https://pokeapi.co/api/v2/pokemon-form/132/"}],"game_indices":[{"game_index":76,"version":{"name":"red","url":"https://pokeapi.co/api/v2/version/1/"}},{"game_index":76,"version":{"name":"blue","url":"https://pokeapi.co/api/v2/version/2/"}},{"game_index":76,"version":{"name":"yellow","url":"https://pokeapi.co/api/v2/version/3/"}},{"game_index":132,"version":{"name":"gold","url":"https://pokeapi.co/api/v2/version/4/"}},{"game_index":132,"version":{"name":"silver","url":"https://pokeapi.co/api/v2/version/5/"}},{"game_index":132,"version":{"name":"crystal","url":"https://pokeapi.co/api/v2/version/6/"}},{"game_index":132,"version":{"name":"ruby","url":"https://pokeapi.co/api/v2/version/7/"}},{"game_index":132,"version":{"name":"sapphire","url":"https://pokeapi.co/api/v2/version/8/"}},{"game_index":132,"version":{"name":"emerald","url":"https://pokeapi.co/api/v2/version/9/"}},{"game_index":132,"version":{"name":"firered","url":"https://pokeapi.co/api/v2/version/10/"}},{"game_index":132,"version":{"name":"leafgreen","url":"https://pokeapi.co/api/v2/version/11/"}},{"game_index":132,"version":{"name":"diamond","url":"https://pokeapi.co/api/v2/version/12/"}},{"game_index":132,"version":{"name":"pearl","url":"https://pokeapi.co/api/v2/version/13/"}},{"game_index":132,"version":{"name":"platinum","url":"https://pokeapi.co/api/v2/version/14/"}},{"game_index":132,"version":{"name":"heartgold","url":"https://pokeapi.co/api/v2/version/15/"}},{"game_index":132,"version":{"name":"soulsilver","url":"https://pokeapi.co/api/v2/version/16/"}},{"game_index":132,"version":{"name":"black","url":"https://pokeapi.co/api/v2/version/17/"}},{"game_index":132,"version":{"name":"white","url":"https://pokeapi.co/api/v2/version/18/"}},{"game_index":132,"version":{"name":"black-2","url":"https://pokeapi.co/api/v2/version/21/"}},{"game_index":132,"version":{"name":"white-2","url":"https://pokeapi.co/api/v2/version/22/"}}],"height":3,"held_items":[{"item":{"name":"metal-powder","url":"https://pokeapi.co/api/v2/item/234/"},"version_details":[{"rarity":5,"version":{"name":"ruby","url":"https://pokeapi.co/api/v2/version/7/"}},{"rarity":5,"version":{"name":"sapphire","url":"https://pokeapi.co/api/v2/version/8/"}},{"rarity":5,"version":{"name":"emerald","url":"https://pokeapi.co/api/v2/version/9/"}},{"rarity":5,"version":{"name":"firered","url":"https://pokeapi.co/api/v2/version/10/"}},{"rarity":5,"version":{"name":"leafgreen","url":"https://pokeapi.co/api/v2/version/11/"}},{"rarity":5,"version":{"name":"diamond","url":"https://pokeapi.co/api/v2/version/12/"}},{"rarity":5,"version":{"name":"pearl","url":"https://pokeapi.co/api/v2/version/13/"}},{"rarity":5,"version":{"name":"platinum","url":"https://pokeapi.co/api/v2/version/14/"}},{"rarity":5,"version":{"name":"heartgold","url":"https://pokeapi.co/api/v2/version/15/"}},{"rarity":5,"version":{"name":"soulsilver","url":"https://pokeapi.co/api/v2/version/16/"}},{"rarity":5,"version":{"name":"black","url":"https://pokeapi.co/api/v2/version/17/"}},{"rarity":5,"version":{"name":"white","url":"https://pokeapi.co/api/v2/version/18/"}},{"rarity":5,"version":{"name":"black-2","url":"https://pokeapi.co/api/v2/version/21/"}},{"rarity":5,"version":{"name":"white-2","url":"https://pokeapi.co/api/v2/version/22/"}},{"rarity":5,"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"rarity":5,"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"rarity":5,"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"rarity":5,"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"rarity":5,"version":{"name":"sun","url":"https://pokeapi.co/api/v2/version/27/"}},{"rarity":5,"version":{"name":"moon","url":"https://pokeapi.co/api/v2/version/28/"}},{"rarity":5,"version":{"name":"ultra-sun","url":"https://pokeapi.co/api/v2/version/29/"}},{"rarity":5,"version":{"name":"ultra-moon","url":"https://pokeapi.co/api/v2/version/30/"}}]},{"item":{"name":"quick-powder","url":"https://pokeapi.co/api/v2/item/251/"},"version_details":[{"rarity":50,"version":{"name":"diamond","url":"https://pokeapi.co/api/v2/version/12/"}},{"rarity":50,"version":{"name":"pearl","url":"https://pokeapi.co/api/v2/version/13/"}},{"rarity":50,"version":{"name":"platinum","url":"https://pokeapi.co/api/v2/version/14/"}},{"rarity":50,"version":{"name":"heartgold","url":"https://pokeapi.co/api/v2/version/15/"}},{"rarity":50,"version":{"name":"soulsilver","url":"https://pokeapi.co/api/v2/version/16/"}},{"rarity":50,"version":{"name":"black","url":"https://pokeapi.co/api/v2/version/17/"}},{"rarity":50,"version":{"name":"white","url":"https://pokeapi.co/api/v2/version/18/"}},{"rarity":50,"version":{"name":"black-2","url":"https://pokeapi.co/api/v2/version/21/"}},{"rarity":50,"version":{"name":"white-2","url":"https://pokeapi.co/api/v2/version/22/"}},{"rarity":50,"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"rarity":50,"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"rarity":50,"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"rarity":50,"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"rarity":50,"version":{"name":"sun","url":"https://pokeapi.co/api/v2/version/27/"}},{"rarity":50,"version":{"name":"moon","url":"https://pokeapi.co/api/v2/version/28/"}},{"rarity":50,"version":{"name":"ultra-sun","url":"https://pokeapi.co/api/v2/version/29/"}},{"rarity":50,"version":{"name":"ultra-moon","url":"https://pokeapi.co/api/v2/version/30/"}}]}],"id":132,"is_default":true,"location_area_encounters":"https://pokeapi.co/api/v2/pokemon/132/encounters","moves":[{"move":{"name":"transform","url":"https://pokeapi.co/api/v2/move/144/"},"version_group_details":[{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"red-blue","url":"https://pokeapi.co/api/v2/version-group/1/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"yellow","url":"https://pokeapi.co/api/v2/version-group/2/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"gold-silver","url":"https://pokeapi.co/api/v2/version-group/3/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"crystal","url":"https://pokeapi.co/api/v2/version-group/4/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"ruby-sapphire","url":"https://pokeapi.co/api/v2/version-group/5/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"emerald","url":"https://pokeapi.co/api/v2/version-group/6/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"firered-leafgreen","url":"https://pokeapi.co/api/v2/version-group/7/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"diamond-pearl","url":"https://pokeapi.co/api/v2/version-group/8/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"platinum","url":"https://pokeapi.co/api/v2/version-group/9/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"heartgold-soulsilver","url":"https://pokeapi.co/api/v2/version-group/10/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"black-white","url":"https://pokeapi.co/api/v2/version-group/11/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"colosseum","url":"https://pokeapi.co/api/v2/version-group/12/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"xd","url":"https://pokeapi.co/api/v2/version-group/13/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"black-2-white-2","url":"https://pokeapi.co/api/v2/version-group/14/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"x-y","url":"https://pokeapi.co/api/v2/version-group/15/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"omega-ruby-alpha-sapphire","url":"https://pokeapi.co/api/v2/version-group/16/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"sun-moon","url":"https://pokeapi.co/api/v2/version-group/17/"}},{"level_learned_at":1,"move_learn_method":{"name":"level-up","url":"https://pokeapi.co/api/v2/move-learn-method/1/"},"version_group":{"name":"ultra-sun-ultra-moon","url":"https://pokeapi.co/api/v2/version-group/18/"}}]}],"name":"ditto","order":203,"species":{"name":"ditto","url":"https://pokeapi.co/api/v2/pokemon-species/132/"},"sprites":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/back/132.png","back_female":null,"back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/back/shiny/132.png","back_shiny_female":null,"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/132.png","front_female":null,"front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/shiny/132.png","front_shiny_female":null,"other":{"dream_world":{"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/132.svg","front_female":null},"official-artwork":{"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/132.png"}},"versions":{"generation-i":{"red-blue":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-i/red-blue/back/132.png","back_gray":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-i/red-blue/back/gray/132.png","front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-i/red-blue/132.png","front_gray":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-i/red-blue/gray/132.png"},"yellow":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-i/yellow/back/132.png","back_gray":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-i/yellow/back/gray/132.png","front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-i/yellow/132.png","front_gray":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-i/yellow/gray/132.png"}},"generation-ii":{"crystal":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/crystal/back/132.png","back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/crystal/back/shiny/132.png","front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/crystal/132.png","front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/crystal/shiny/132.png"},"gold":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/gold/back/132.png","back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/gold/back/shiny/132.png","front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/gold/132.png","front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/gold/shiny/132.png"},"silver":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/silver/back/132.png","back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/silver/back/shiny/132.png","front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/silver/132.png","front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-ii/silver/shiny/132.png"}},"generation-iii":{"emerald":{"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/emerald/132.png","front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/emerald/shiny/132.png"},"firered-leafgreen":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/firered-leafgreen/back/132.png","back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/firered-leafgreen/back/shiny/132.png","front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/firered-leafgreen/132.png","front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/firered-leafgreen/shiny/132.png"},"ruby-sapphire":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/ruby-sapphire/back/132.png","back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/ruby-sapphire/back/shiny/132.png","front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/ruby-sapphire/132.png","front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iii/ruby-sapphire/shiny/132.png"}},"generation-iv":{"diamond-pearl":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/diamond-pearl/back/132.png","back_female":null,"back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/diamond-pearl/back/shiny/132.png","back_shiny_female":null,"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/diamond-pearl/132.png","front_female":null,"front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/diamond-pearl/shiny/132.png","front_shiny_female":null},"heartgold-soulsilver":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/heartgold-soulsilver/back/132.png","back_female":null,"back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/heartgold-soulsilver/back/shiny/132.png","back_shiny_female":null,"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/heartgold-soulsilver/132.png","front_female":null,"front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/heartgold-soulsilver/shiny/132.png","front_shiny_female":null},"platinum":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/platinum/back/132.png","back_female":null,"back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/platinum/back/shiny/132.png","back_shiny_female":null,"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/platinum/132.png","front_female":null,"front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-iv/platinum/shiny/132.png","front_shiny_female":null}},"generation-v":{"black-white":{"animated":{"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-v/black-white/animated/back/132.gif","back_female":null,"back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-v/black-white/animated/back/shiny/132.gif","back_shiny_female":null,"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-v/black-white/animated/132.gif","front_female":null,"front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-v/black-white/animated/shiny/132.gif","front_shiny_female":null},"back_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-v/black-white/back/132.png","back_female":null,"back_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-v/black-white/back/shiny/132.png","back_shiny_female":null,"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-v/black-white/132.png","front_female":null,"front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-v/black-white/shiny/132.png","front_shiny_female":null}},"generation-vi":{"omegaruby-alphasapphire":{"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-vi/omegaruby-alphasapphire/132.png","front_female":null,"front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-vi/omegaruby-alphasapphire/shiny/132.png","front_shiny_female":null},"x-y":{"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-vi/x-y/132.png","front_female":null,"front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-vi/x-y/shiny/132.png","front_shiny_female":null}},"generation-vii":{"icons":{"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-vii/icons/132.png","front_female":null},"ultra-sun-ultra-moon":{"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-vii/ultra-sun-ultra-moon/132.png","front_female":null,"front_shiny":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-vii/ultra-sun-ultra-moon/shiny/132.png","front_shiny_female":null}},"generation-viii":{"icons":{"front_default":"https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/versions/generation-viii/icons/132.png","front_female":null}}}},"stats":[{"base_stat":48,"effort":1,"stat":{"name":"hp","url":"https://pokeapi.co/api/v2/stat/1/"}},{"base_stat":48,"effort":0,"stat":{"name":"attack","url":"https://pokeapi.co/api/v2/stat/2/"}},{"base_stat":48,"effort":0,"stat":{"name":"defense","url":"https://pokeapi.co/api/v2/stat/3/"}},{"base_stat":48,"effort":0,"stat":{"name":"special-attack","url":"https://pokeapi.co/api/v2/stat/4/"}},{"base_stat":48,"effort":0,"stat":{"name":"special-defense","url":"https://pokeapi.co/api/v2/stat/5/"}},{"base_stat":48,"effort":0,"stat":{"name":"speed","url":"https://pokeapi.co/api/v2/stat/6/"}}],"types":[{"slot":1,"type":{"name":"normal","url":"https://pokeapi.co/api/v2/type/1/"}}],"weight":40}"""
+
+        //Response of pokemon-species/charizard or pokemon-species/6
         const val DETAIL_CHARIZARD = """{"base_happiness":70,"capture_rate":45,"color":{"name":"red","url":"https://pokeapi.co/api/v2/pokemon-color/8/"},"egg_groups":[{"name":"monster","url":"https://pokeapi.co/api/v2/egg-group/1/"},{"name":"dragon","url":"https://pokeapi.co/api/v2/egg-group/14/"}],"evolution_chain":{"url":"https://pokeapi.co/api/v2/evolution-chain/2/"},"evolves_from_species":{"name":"charmeleon","url":"https://pokeapi.co/api/v2/pokemon-species/5/"},"flavor_text_entries":[{"flavor_text":"Spits fire that\nis hot enough to\nmelt boulders.\fKnown to cause\nforest fires\nunintentionally.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"red","url":"https://pokeapi.co/api/v2/version/1/"}},{"flavor_text":"Spits fire that\nis hot enough to\nmelt boulders.\fKnown to cause\nforest fires\nunintentionally.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"blue","url":"https://pokeapi.co/api/v2/version/2/"}},{"flavor_text":"When expelling a\nblast of super\nhot fire, the red\fflame at the tip\nof its tail burns\nmore intensely.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"yellow","url":"https://pokeapi.co/api/v2/version/3/"}},{"flavor_text":"If CHARIZARD be­\ncomes furious, the\nflame at the tip\fof its tail flares\nup in a whitish-\nblue color.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"gold","url":"https://pokeapi.co/api/v2/version/4/"}},{"flavor_text":"Breathing intense,\nhot flames, it can\nmelt almost any­\fthing. Its breath\ninflicts terrible\npain on enemies.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"silver","url":"https://pokeapi.co/api/v2/version/5/"}},{"flavor_text":"It uses its wings\nto fly high. The\ntemperature of its\ffire increases as\nit gains exper­\nience in battle.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"crystal","url":"https://pokeapi.co/api/v2/version/6/"}},{"flavor_text":"CHARIZARD flies around the sky in\nsearch of powerful opponents.\nIt breathes fire of such great heat\fthat it melts anything. However, it\nnever turns its fiery breath on any\nopponent weaker than itself.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"ruby","url":"https://pokeapi.co/api/v2/version/7/"}},{"flavor_text":"CHARIZARD flies around the sky in\nsearch of powerful opponents.\nIt breathes fire of such great heat\fthat it melts anything. However, it\nnever turns its fiery breath on any\nopponent weaker than itself.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"sapphire","url":"https://pokeapi.co/api/v2/version/8/"}},{"flavor_text":"A CHARIZARD flies about in search of\nstrong opponents. It breathes intense\nflames that can melt any material. However,\nit will never torch a weaker foe.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"emerald","url":"https://pokeapi.co/api/v2/version/9/"}},{"flavor_text":"Its wings can carry this POKéMON close to\nan altitude of 4,600 feet. It blows out\nfire at very high temperatures.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"firered","url":"https://pokeapi.co/api/v2/version/10/"}},{"flavor_text":"It spits fire that is hot enough to melt\nboulders. It may cause forest fires by\nblowing flames.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"leafgreen","url":"https://pokeapi.co/api/v2/version/11/"}},{"flavor_text":"It is said that CHARIZARD’s fire\nburns hotter if it has\nexperienced harsh battles.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"diamond","url":"https://pokeapi.co/api/v2/version/12/"}},{"flavor_text":"It is said that CHARIZARD’s fire\nburns hotter if it has\nexperienced harsh battles.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"pearl","url":"https://pokeapi.co/api/v2/version/13/"}},{"flavor_text":"It is said that CHARIZARD’s fire\nburns hotter if it has\nexperienced harsh battles.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"platinum","url":"https://pokeapi.co/api/v2/version/14/"}},{"flavor_text":"If CHARIZARD becomes furious,\nthe flame at the tip of its tail flares\nup in a light blue shade.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"heartgold","url":"https://pokeapi.co/api/v2/version/15/"}},{"flavor_text":"Breathing intense, hot flames, it can\nmelt almost anything. Its breath\ninflicts terrible pain on enemies.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"soulsilver","url":"https://pokeapi.co/api/v2/version/16/"}},{"flavor_text":"On raconte que la flamme du\nDracaufeu s’intensifie après\nun combat difficile.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"black","url":"https://pokeapi.co/api/v2/version/17/"}},{"flavor_text":"It is said that Charizard’s fire\nburns hotter if it has\nexperienced harsh battles.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"black","url":"https://pokeapi.co/api/v2/version/17/"}},{"flavor_text":"On raconte que la flamme du\nDracaufeu s’intensifie après\nun combat difficile.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"white","url":"https://pokeapi.co/api/v2/version/18/"}},{"flavor_text":"It is said that Charizard’s fire\nburns hotter if it has\nexperienced harsh battles.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"white","url":"https://pokeapi.co/api/v2/version/18/"}},{"flavor_text":"It is said that Charizard’s fire\nburns hotter if it has\nexperienced harsh battles.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"black-2","url":"https://pokeapi.co/api/v2/version/21/"}},{"flavor_text":"It is said that Charizard’s fire\nburns hotter if it has\nexperienced harsh battles.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"white-2","url":"https://pokeapi.co/api/v2/version/22/"}},{"flavor_text":"くちから　しゃくねつの　ほのおを\nはきだすとき　しっぽの　さきは\nより　あかく　はげしく　もえあがる。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"flavor_text":"입에서 작렬하는 불꽃을\n토해낼 때 꼬리의 끝이\n더욱 붉고 격렬하게 타오른다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"flavor_text":"Quand il crache son souffle brûlant, la flamme au bout\nde sa queue s’embrase.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"flavor_text":"Wenn dieses Pokémon einen Strahl glühenden\nFeuers speit, leuchtet seine Schwanzspitze auf.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"flavor_text":"Cuando lanza una descarga de fuego supercaliente, la\nroja llama de su cola brilla más intensamente.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"flavor_text":"Quando emette le sue lingue di fuoco, la fiamma\nrossa sulla punta della coda brucia più intensamente.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"flavor_text":"When expelling a blast of superhot fire,\nthe red flame at the tip of its tail burns\nmore intensely.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"flavor_text":"口から　灼熱の　炎を　吐き出すとき\n尻尾の　先は\nより　赤く　激しく　燃え上がる。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"x","url":"https://pokeapi.co/api/v2/version/23/"}},{"flavor_text":"ちじょう　１４００メートル　まで\nハネを　つかって　とぶことができる。\nこうねつの　ほのおを　はく。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"flavor_text":"지상 1400m까지\n날개를 사용해 날 수 있다.\n고열의 불꽃을 내뿜는다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"flavor_text":"Ses ailes peuvent le faire voler à plus de 1 400 m\nd’altitude. Ce Pokémon crache du feu à des\ntempératures très élevées.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"flavor_text":"Dieses Pokémon kann mit seinen Flügeln\neine Höhe von bis zu 1 400 m erreichen.\nEs spuckt sehr heißes Feuer.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"flavor_text":"Con las alas que tiene puede alcanzar una altura de\ncasi 1400 m. Suele escupir fuego por la boca.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"flavor_text":"Grazie alle possenti ali può volare fino a 1400 m\nd’altezza. Sputa fuoco a una temperatura\nimpressionante.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"flavor_text":"Its wings can carry this Pokémon close to an\naltitude of 4,600 feet. It blows out fire at very\nhigh temperatures.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"flavor_text":"地上　１４００メートルまで\n羽を　使って　飛ぶことができる。\n高熱の　炎を　吐く。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"y","url":"https://pokeapi.co/api/v2/version/24/"}},{"flavor_text":"つよい　あいてを　もとめて　そらを　とびまわる。\nなんでも　とかして　しまう　こうねつの　ほのおを\nじぶんより　よわいものに　むけることは　しない。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"flavor_text":"강한 상대를 찾아 하늘을 날아다닌다.\n무엇이든 다 녹여버리는 고열의 불꽃을\n자신보다 약한 자에게 들이대지 않는다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"flavor_text":"Dracaufeu parcourt les cieux pour trouver des adversaires\nà sa mesure. Il crache de puissantes flammes capables de\nfaire fondre n’importe quoi. Mais il ne dirige jamais son souffle\ndestructeur vers un ennemi plus faible.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"flavor_text":"Glurak fliegt durch die Lüfte, um starke Gegner aufzuspüren.\nSein heißer Feueratem bringt alles zum Schmelzen. Aber es\nrichtet seinen Feueratem nie auf schwächere Gegner.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"flavor_text":"Charizard se dedica a volar por los cielos en busca de\noponentes fuertes. Echa fuego por la boca y es capaz de\nderretir cualquier cosa. No obstante, si su rival es más débil\nque él, no usará este ataque.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"flavor_text":"Charizard solca i cieli in cerca di nemici molto forti.\nRiesce a emettere fiammate di un calore tale da fondere\nogni cosa. Tuttavia, non rivolge mai le sue micidiali lingue\ndi fuoco contro avversari più deboli di lui.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"flavor_text":"Charizard flies around the sky in search of powerful opponents.\nIt breathes fire of such great heat that it melts anything.\nHowever, it never turns its fiery breath on any opponent\nweaker than itself.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"flavor_text":"強い　相手を　求めて　空を　飛び回る。\nなんでも　溶かして　しまう　高熱の　炎を\n自分より　弱いものに　向けることは　しない。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"omega-ruby","url":"https://pokeapi.co/api/v2/version/25/"}},{"flavor_text":"つよい　あいてを　もとめて　そらを　とびまわる。\nなんでも　とかして　しまう　こうねつの　ほのおを\nじぶんより　よわいものに　むけることは　しない。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"flavor_text":"강한 상대를 찾아 하늘을 날아다닌다.\n무엇이든 다 녹여버리는 고열의 불꽃을\n자신보다 약한 자에게 들이대지 않는다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"flavor_text":"Dracaufeu parcourt les cieux pour trouver des adversaires\nà sa mesure. Il crache de puissantes flammes capables de\nfaire fondre n’importe quoi. Mais il ne dirige jamais son souffle\ndestructeur vers un ennemi plus faible.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"flavor_text":"Glurak fliegt durch die Lüfte, um starke Gegner aufzuspüren.\nSein heißer Feueratem bringt alles zum Schmelzen. Aber es\nrichtet seinen Feueratem nie gegen schwächere Gegner.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"flavor_text":"Charizard se dedica a volar por los cielos en busca de\noponentes fuertes. Echa fuego por la boca y es capaz de\nderretir cualquier cosa. No obstante, si su rival es más débil\nque él, no usará este ataque.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"flavor_text":"Charizard solca i cieli in cerca di nemici molto forti.\nRiesce a emettere fiammate di un calore tale da fondere\nogni cosa. Tuttavia, non rivolge mai le sue micidiali lingue\ndi fuoco contro avversari più deboli di lui.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"flavor_text":"Charizard flies around the sky in search of powerful opponents.\nIt breathes fire of such great heat that it melts anything.\nHowever, it never turns its fiery breath on any opponent\nweaker than itself.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"flavor_text":"強い　相手を　求めて　空を　飛び回る。\nなんでも　溶かして　しまう　高熱の　炎を\n自分より　弱いものに　向けることは　しない。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"alpha-sapphire","url":"https://pokeapi.co/api/v2/version/26/"}},{"flavor_text":"くちから　しゃくねつの　ほのおを\nはきだすとき　シッポのさきは\nより　あかく　はげしく　もえあがる。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"ぜんしんに　みなぎる　なみはずれた\nパワーが　からだを　くろく　そめあげ\nあおい　ほのおを　メラメラ　もやす。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"トレーナーとの　きずなが　パワーの\nみなもと。　ジェットきを　しのぐ\nひこう　のうりょくを　ほこる。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"온몸에 넘치는 엄청난 파워가\n몸을 검게 물들이고\n푸른 불꽃을 이글이글 태운다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"입에서 작렬하는 불꽃을\n토해낼 때 꼬리의 끝이\n더욱 붉고 격렬하게 타오른다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"트레이너와의 유대가 힘의\n원천이다. 제트기를 능가하는\n비행 능력을 자랑한다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"全身湧出超乎尋常的力量，\n將身體染成了黑色，同時\n熊熊燃燒著藍色的火焰。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"從口中噴出熾熱的火焰時，\n尾巴尖端的紅色火焰\n會燃燒得更加激烈。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"與訓練家的牽絆是牠力量的\n泉源。引以為傲的飛行能力\n甚至凌駕於噴射機之上。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"Il tire sa puissance du lien qui l’unit à son\nDresseur. Sa capacité à voler plus rapidement\nqu’un avion à réaction fait toute sa fierté.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"La puissance incommensurable parcourant son\ncorps est à l’origine des flammes bleues qu’il\ndéploie et de la teinte sombre de sa peau.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"Quand il crache son souffle brûlant, la flamme\nau bout de sa queue s’embrase.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"Aus der Bindung zu seinem Trainer schöpft es\nseine Kraft. Seine Fluggeschwindigkeit übertrifft\nsogar die eines Düsenjets.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"Es wird durch und durch von einer mächtigen\nKraft durchströmt, die seinen Körper schwarz\nfärbt und seine Flammen blau auflodern lässt.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"Wenn es einen Strahl glühenden Feuers speit,\nlodert die rote Flamme an seiner Schwanzspitze\nnoch intensiver.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"Cuando exhala fuego abrasador, la llama de la\npunta de la cola se aviva y adquiere un intenso\ncolor rojo.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"El vínculo que lo une a su Entrenador es la\nfuente de su poder. Su capacidad de vuelo\nsupera la de un avión a reacción.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"La extraordinaria energía que fluye por su\ncuerpo le ha ennegrecido la piel y provocado\nla aparición de intensas llamas azules.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"Il legame con il suo Allenatore è la fonte del suo\npotere. Vanta una velocità di volo superiore\na quella di un aereo a reazione.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"La straordinaria potenza che pervade il suo\ncorpo lo rende più scuro e genera roventi\nfiamme che si colorano di blu.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"Quando emette lingue di fuoco roventi\ndalla bocca, la fiamma rossa sulla punta\ndella coda brucia più intensamente.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"Its bond with its Trainer is the source of\nits power. It boasts speed and maneuverability\ngreater than that of a jet fighter.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"The overwhelming power that fills its entire body\ncauses it to turn black and creates intense\nblue flames.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"When this Pokémon expels a blast of superhot\nfire, the red flame at the tip of its tail burns\nmore intensely.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"トレーナーとの　絆が　パワーの　源。\nジェット機を　しのぐ\n飛行能力を　誇る。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"全身に　みなぎる　並はずれた\nパワーが　体を　黒く　染め上げ\n青い　炎を　メラメラ　燃やす。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"口から　灼熱の　炎を\n吐き出すとき　シッポの先は\nより　赤く　激しく　燃えあがる。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"与训练家的牵绊是它的力量\n之源。它引以为傲的飞行能力\n完全凌驾于喷气式飞机之上。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"从口中喷出灼热的火焰时，\n尾巴尖端的红色火焰\n会燃烧得更加猛烈。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"全身充斥着非比寻常的力量，\n这种力量染黑了它的身体，\n燃烧出炽热的蓝色火焰。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"lets-go-pikachu","url":"https://pokeapi.co/api/v2/version/31/"}},{"flavor_text":"くちから　しゃくねつの　ほのおを\nはきだすとき　シッポのさきは\nより　あかく　はげしく　もえあがる。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"ぜんしんに　みなぎる　なみはずれた\nパワーが　からだを　くろく　そめあげ\nあおい　ほのおを　メラメラ　もやす。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"トレーナーとの　きずなが　パワーの\nみなもと。　ジェットきを　しのぐ\nひこう　のうりょくを　ほこる。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"온몸에 넘치는 엄청난 파워가\n몸을 검게 물들이고\n푸른 불꽃을 이글이글 태운다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"입에서 작렬하는 불꽃을\n토해낼 때 꼬리의 끝이\n더욱 붉고 격렬하게 타오른다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"트레이너와의 유대가 힘의\n원천이다. 제트기를 능가하는\n비행 능력을 자랑한다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"全身湧出超乎尋常的力量，\n將身體染成了黑色，同時\n熊熊燃燒著藍色的火焰。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"從口中噴出熾熱的火焰時，\n尾巴尖端的紅色火焰\n會燃燒得更加激烈。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"與訓練家的牽絆是牠力量的\n泉源。引以為傲的飛行能力\n甚至凌駕於噴射機之上。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"Il tire sa puissance du lien qui l’unit à son\nDresseur. Sa capacité à voler plus rapidement\nqu’un avion à réaction fait toute sa fierté.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"La puissance incommensurable parcourant son\ncorps est à l’origine des flammes bleues qu’il\ndéploie et de la teinte sombre de sa peau.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"Quand il crache son souffle brûlant, la flamme\nau bout de sa queue s’embrase.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"Aus der Bindung zu seinem Trainer schöpft es\nseine Kraft. Seine Fluggeschwindigkeit übertrifft\nsogar die eines Düsenjets.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"Es wird durch und durch von einer mächtigen\nKraft durchströmt, die seinen Körper schwarz\nfärbt und seine Flammen blau auflodern lässt.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"Wenn es einen Strahl glühenden Feuers speit,\nlodert die rote Flamme an seiner Schwanzspitze\nnoch intensiver.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"Cuando exhala fuego abrasador, la llama de la\npunta de la cola se aviva y adquiere un intenso\ncolor rojo.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"El vínculo que lo une a su Entrenador es la\nfuente de su poder. Su capacidad de vuelo\nsupera la de un avión a reacción.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"La extraordinaria energía que fluye por su\ncuerpo le ha ennegrecido la piel y provocado\nla aparición de intensas llamas azules.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"Il legame con il suo Allenatore è la fonte del suo\npotere. Vanta una velocità di volo superiore\na quella di un aereo a reazione.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"La straordinaria potenza che pervade il suo\ncorpo lo rende più scuro e genera roventi\nfiamme che si colorano di blu.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"Quando emette lingue di fuoco roventi\ndalla bocca, la fiamma rossa sulla punta\ndella coda brucia più intensamente.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"Its bond with its Trainer is the source of\nits power. It boasts speed and maneuverability\ngreater than that of a jet fighter.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"The overwhelming power that fills its entire body\ncauses it to turn black and creates intense\nblue flames.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"When this Pokémon expels a blast of superhot\nfire, the red flame at the tip of its tail burns\nmore intensely.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"トレーナーとの　絆が　パワーの　源。\nジェット機を　しのぐ\n飛行能力を　誇る。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"全身に　みなぎる　並はずれた\nパワーが　体を　黒く　染め上げ\n青い　炎を　メラメラ　燃やす。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"口から　灼熱の　炎を\n吐き出すとき　シッポの先は\nより　赤く　激しく　燃えあがる。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"与训练家的牵绊是它的力量\n之源。它引以为傲的飞行能力\n完全凌驾于喷气式飞机之上。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"从口中喷出灼热的火焰时，\n尾巴尖端的红色火焰\n会燃烧得更加猛烈。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"全身充斥着非比寻常的力量，\n这种力量染黑了它的身体，\n燃烧出炽热的蓝色火焰。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"lets-go-eevee","url":"https://pokeapi.co/api/v2/version/32/"}},{"flavor_text":"がんせきも　やけるような\nしゃくねつの　ほのおを　はいて\nやまかじを　おこすことが　ある。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"キョダイマックスの　パワーによって\nほのおの　つばさを　もつ　きょだいな\nリザードンの　すがたが　つくられた。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"거다이맥스의 파워에 의해\n불꽃의 날개를 가진 거대한\n리자몽의 모습이 만들어졌다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"암석도 태워버릴 정도로\n작열하는 화염을 뿜어\n산불을 일으킬 때가 있다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"會噴出彷彿連岩石\n都能燒焦的灼熱火焰。\n有時會引發森林火災。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"超極巨化之力不僅讓\n噴火龍變得更加巨大，\n還讓牠獲得了熊熊燃燒的翅膀。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"Grâce au phénomène Gigamax, il a pris la forme\nd’un Dracaufeu géant doté d’ailes de feu.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"Son souffle brûlant peut faire fondre la roche.\nIl est parfois la cause d’incendies de forêt.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"Dieses Pokémon kann mit seinem Feueratem\nFelsen schmelzen. Es verursacht ab und zu\nWaldbrände.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"Durch die Gigadynamax-Energie ist dieses\ngigantische Glurak entstanden, das über Flügel\naus Flammen verfügt.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"Escupe un fuego tan caliente que funde las rocas.\nCausa incendios forestales sin querer.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"La energía del fenómeno Gigamax ha dotado a\neste Charizard de alas flamígeras y un enorme\ntamaño.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"L’energia del fenomeno Gigamax lo ha\ntrasformato in un Charizard gigante dalle\nali di fuoco.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"Sputa fiamme incandescenti in grado di fondere\nle rocce. A volte causa incendi boschivi.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"It spits fire that is hot enough to melt boulders.\nIt may cause forest fires by blowing flames.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"This colossal, flame-winged figure of a Charizard\nwas brought about by Gigantamax energy.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"キョダイマックスの　パワーによって\n炎の　翼を　持つ　巨大な\nリザードンの　姿が　つくられた。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"岩石も　焼けるような\n灼熱の　炎を　吐いて\n山火事を　起こすことが　ある。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"能够喷出猛烈的火焰，\n仿佛连岩石都能烤焦。\n有时会引发森林火灾。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"超极巨化之力不仅让喷火龙\n变得更加巨大，还让它\n获得了熊熊燃烧的翅膀。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"sword","url":"https://pokeapi.co/api/v2/version/33/"}},{"flavor_text":"せっし２０００どの　ほのおが\nたいないで　うずまく。　ほえると\nかりょくが　さらに　あがる。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"ちじょう　１４００メートル　まで\nハネを　つかって　とぶことができる。\nこうねつの　ほのおを　はく。","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"섭씨 2000도의 불꽃이\n체내에서 소용돌이친다.\n울부짖으면 화력이 더 올라간다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"지상 1400m까지\n날개를 사용해 날 수 있다.\n고열의 불꽃을 내뿜는다.","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"攝氏２０００度的火焰在體內\n形成了旋渦。當牠咆哮的時候，\n火力還會進一步提升。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"能用翅膀飛到距離地面\n１４００公尺的高度。\n會吐出高溫的火焰。","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"Des flammes, dont la température atteint\nles 2 000 °C, dansent à l’intérieur de son corps.\nQuand il rugit, leur puissance augmente.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"Ses ailes lui permettent de voler à plus de\n1 400 m d’altitude. Ce Pokémon crache du feu\nà des températures très élevées.","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"Dieses Pokémon kann mit seinen Flügeln eine\nHöhe von bis zu 1 400 m erreichen. Es spuckt\nsehr heißes Feuer.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"In seinem Inneren wirbeln 2 000 ºC heiße\nFlammen umher. Wenn es brüllt, steigt die\nFeuerkraft noch weiter an.","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"Las llamas del interior de su cuerpo forman un\ntorbellino que alcanza los 2000 °C. Al rugir\naumenta todavía más su energía térmica.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"Sus potentes alas le permiten volar a una altura\nde 1400 m. Escupe llamaradas que llegan a\nalcanzar temperaturas elevadísimas.","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"All’interno del suo corpo turbina il fuoco a una\ntemperatura di 2.000 ºC. Quando ruggisce, la\nsua energia termica aumenta ulteriormente.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"Grazie alle possenti ali può volare fino a\n1.400 m d’altezza. Le fiamme che sputa\npossono raggiungere temperature altissime.","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"Its wings can carry this Pokémon close to an\naltitude of 4,600 feet. It blows out fire at very\nhigh temperatures.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"The flame inside its body burns hotter than\n3,600 degrees Fahrenheit. When Charizard\nroars, that temperature climbs even higher.","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"地上　１４００メートル　まで\n羽を　使って　飛ぶことができる。\n高熱の　炎を　吐く。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"摂氏２０００度の　炎が\n体内で　渦巻く。　吠えると\n火力が　さらに　上がる。","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"能用翅膀飞到距地面\n１４００米的高度。\n会吐出高温火焰。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}},{"flavor_text":"２０００度的火焰在体内形成\n了漩涡。当它咆哮的时候，\n火力还会进一步上升。","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"version":{"name":"shield","url":"https://pokeapi.co/api/v2/version/34/"}}],"form_descriptions":[],"forms_switchable":true,"gender_rate":1,"genera":[{"genus":"かえんポケモン","language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"}},{"genus":"화염포켓몬","language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"}},{"genus":"火焰寶可夢","language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"}},{"genus":"Pokémon Flamme","language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"}},{"genus":"Flamme","language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"}},{"genus":"Pokémon Llama","language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"}},{"genus":"Pokémon Fiamma","language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"}},{"genus":"Flame Pokémon","language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"}},{"genus":"かえんポケモン","language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"}},{"genus":"火焰宝可梦","language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"}}],"generation":{"name":"generation-i","url":"https://pokeapi.co/api/v2/generation/1/"},"growth_rate":{"name":"medium-slow","url":"https://pokeapi.co/api/v2/growth-rate/4/"},"habitat":{"name":"mountain","url":"https://pokeapi.co/api/v2/pokemon-habitat/4/"},"has_gender_differences":false,"hatch_counter":20,"id":6,"is_baby":false,"is_legendary":false,"is_mythical":false,"name":"charizard","names":[{"language":{"name":"ja-Hrkt","url":"https://pokeapi.co/api/v2/language/1/"},"name":"リザードン"},{"language":{"name":"roomaji","url":"https://pokeapi.co/api/v2/language/2/"},"name":"Lizardon"},{"language":{"name":"ko","url":"https://pokeapi.co/api/v2/language/3/"},"name":"리자몽"},{"language":{"name":"zh-Hant","url":"https://pokeapi.co/api/v2/language/4/"},"name":"噴火龍"},{"language":{"name":"fr","url":"https://pokeapi.co/api/v2/language/5/"},"name":"Dracaufeu"},{"language":{"name":"de","url":"https://pokeapi.co/api/v2/language/6/"},"name":"Glurak"},{"language":{"name":"es","url":"https://pokeapi.co/api/v2/language/7/"},"name":"Charizard"},{"language":{"name":"it","url":"https://pokeapi.co/api/v2/language/8/"},"name":"Charizard"},{"language":{"name":"en","url":"https://pokeapi.co/api/v2/language/9/"},"name":"Charizard"},{"language":{"name":"ja","url":"https://pokeapi.co/api/v2/language/11/"},"name":"リザードン"},{"language":{"name":"zh-Hans","url":"https://pokeapi.co/api/v2/language/12/"},"name":"喷火龙"}],"order":6,"pal_park_encounters":[{"area":{"name":"field","url":"https://pokeapi.co/api/v2/pal-park-area/2/"},"base_score":90,"rate":3}],"pokedex_numbers":[{"entry_number":6,"pokedex":{"name":"national","url":"https://pokeapi.co/api/v2/pokedex/1/"}},{"entry_number":6,"pokedex":{"name":"kanto","url":"https://pokeapi.co/api/v2/pokedex/2/"}},{"entry_number":231,"pokedex":{"name":"original-johto","url":"https://pokeapi.co/api/v2/pokedex/3/"}},{"entry_number":236,"pokedex":{"name":"updated-johto","url":"https://pokeapi.co/api/v2/pokedex/7/"}},{"entry_number":111,"pokedex":{"name":"conquest-gallery","url":"https://pokeapi.co/api/v2/pokedex/11/"}},{"entry_number":85,"pokedex":{"name":"kalos-central","url":"https://pokeapi.co/api/v2/pokedex/12/"}},{"entry_number":380,"pokedex":{"name":"galar","url":"https://pokeapi.co/api/v2/pokedex/27/"}},{"entry_number":6,"pokedex":{"name":"updated-kanto","url":"https://pokeapi.co/api/v2/pokedex/26/"}}],"shape":{"name":"upright","url":"https://pokeapi.co/api/v2/pokemon-shape/6/"},"varieties":[{"is_default":true,"pokemon":{"name":"charizard","url":"https://pokeapi.co/api/v2/pokemon/6/"}},{"is_default":false,"pokemon":{"name":"charizard-mega-x","url":"https://pokeapi.co/api/v2/pokemon/10034/"}},{"is_default":false,"pokemon":{"name":"charizard-mega-y","url":"https://pokeapi.co/api/v2/pokemon/10035/"}},{"is_default":false,"pokemon":{"name":"charizard-gmax","url":"https://pokeapi.co/api/v2/pokemon/10187/"}}]}"""
     }
 }
